:hammer: refactor: Refactor Message prevent rerendering
